Move all tournament tests to using placeholder data rather than reading from bracket
