+++ osu.Game.Tournament.Tests/Screens/TestSceneMapPoolScreen.cs
@@ -12,7 +12,7 @@ using osu.Game.Tournament.Screens.MapPool;
 
 namespace osu.Game.Tournament.Tests.Screens
 {
-    public class TestSceneMapPoolScreen : LadderTestScene
+    public class TestSceneMapPoolScreen : TournamentTestScene
     {
         private MapPoolScreen screen;
