+++ osu.Game.Tournament.Tests/Screens/TestSceneSeedingScreen.cs
@@ -8,7 +8,7 @@ using osu.Game.Tournament.Screens.TeamIntro;
 
 namespace osu.Game.Tournament.Tests.Screens
 {
-    public class TestSceneSeedingScreen : LadderTestScene
+    public class TestSceneSeedingScreen : TournamentTestScene
     {
         [Cached]
         private readonly LadderInfo ladder = new LadderInfo();
